+++ reqtls/src/export/cipher.rs
@@ -10,13 +10,13 @@ pub extern "C" fn Cipher_new(ct: CipherType) -> *mut Cipher {
 
 #[unsafe(no_mangle)]
 #[allow(non_snake_case)]
-pub extern "C" fn Cipher_set_secret_key(cipher: *mut Cipher, key: *const u8, key_len: u32, iv: *const u8, iv_len: u32) -> i32 {
+pub extern "C" fn Cipher_set_secret_key(cipher: *mut Cipher, key: *const u8, key_len: usize, iv: *const u8, iv_len: usize) -> i32 {
     || -> RlsResult<i32>{
         let cipher = unsafe { cipher.as_mut() }.ok_or(RlsError::NullPtr)?;
-        let key = unsafe { slice::from_raw_parts(key, key_len as usize) };
+        let key = unsafe { slice::from_raw_parts(key, key_len) };
         let iv = match iv.is_null() {
             true => None,
-            false => Some(unsafe { slice::from_raw_parts(iv, iv_len as usize) }),
+            false => Some(unsafe { slice::from_raw_parts(iv, iv_len) }),
         };
         cipher.set_secret_key(key, iv);
         Ok(0)
@@ -25,14 +25,14 @@ pub extern "C" fn Cipher_set_secret_key(cipher: *mut Cipher, key: *const u8, key
 
 #[unsafe(no_mangle)]
 #[allow(non_snake_case)]
-pub extern "C" fn Cipher_encrypt(cipher: *mut Cipher, ct: *const u8, ct_len: u32, out: *mut *mut u8, out_len: &mut u32) -> i32 {
+pub extern "C" fn Cipher_encrypt(cipher: *mut Cipher, ct: *const u8, ct_len: usize, out: *mut *mut u8, out_len: &mut usize) -> i32 {
     || -> RlsResult<i32>{
         let cipher = unsafe { cipher.as_mut() }.ok_or(RlsError::NullPtr)?;
-        let data = unsafe { slice::from_raw_parts(ct, ct_len as usize) };
+        let data = unsafe { slice::from_raw_parts(ct, ct_len) };
         let mut en_bs = cipher.encrypt(data)?;
         unsafe {
             *out = en_bs.as_mut_ptr();
-            *out_len = en_bs.len() as u32
+            *out_len = en_bs.len()
         };
         mem::forget(en_bs);
         Ok(0)
@@ -41,14 +41,14 @@ pub extern "C" fn Cipher_encrypt(cipher: *mut Cipher, ct: *const u8, ct_len: u32
 
 #[unsafe(no_mangle)]
 #[allow(non_snake_case)]
-pub extern "C" fn Cipher_decrypt(cipher: *mut Cipher, ct: *const u8, ct_len: u32, out: *mut *mut u8, out_len: &mut u32) -> i32 {
+pub extern "C" fn Cipher_decrypt(cipher: *mut Cipher, ct: *const u8, ct_len: usize, out: *mut *mut u8, out_len: &mut usize) -> i32 {
     || -> RlsResult<i32>{
         let cipher = unsafe { cipher.as_mut() }.ok_or(RlsError::NullPtr)?;
-        let data = unsafe { slice::from_raw_parts(ct, ct_len as usize) };
+        let data = unsafe { slice::from_raw_parts(ct, ct_len) };
         let mut en_bs = cipher.decrypt(data)?;
         unsafe {
             *out = en_bs.as_mut_ptr();
-            *out_len = en_bs.len() as u32
+            *out_len = en_bs.len()
         };
         mem::forget(en_bs);
         Ok(0)
